Set working directory for remote rake tasks
Fly.io does support running of rake tasks remotely, but this requires some setup. This commit (and the previous commit) are changes to the dockerfile generated by commands provided by Fly.io to enable runnning interactive consoles as a Rails user, and to set the working directory which seems to be necessary for rake task execution.